add templating to dsl

+++ src/cli/src/build/structurizr.ts
@@ -1,10 +1,17 @@
 import { Build } from "../core/build.js";
+import { matchAll } from "../core/regExp.js";
 import { FileSystem } from "../infrastructure/file-system.js";
 import { Logger } from "../infrastructure/logger.js";
 import { ServiceProvider } from "../infrastructure/service-provider.js";
 import { PlantUml } from "../tools/plant-uml.js";
 import { Structurizr } from "../tools/structurizr.js";
 import { PlantUmlBuild } from "./plant-uml.js";
+
+const templateBeginRegexp = /\s*\/\/\s*@template-begin\s*/gm;
+const templateRenderRegexp = /\/\/\s*@render\s*[^\n]+/gm;
+const renderParametersRegexp = /(\w+)\s*=\s*\"([^"]*)\"/gm;
+const templateBodyRegexp = /\/\/\s*@template-body\n/m;
+const templateEndRegexp = /\/\/\s*@template-end/m;
 
 export namespace StructurizrBuild {
     export function register(serviceProvider: ServiceProvider.IServiceProvider) {
@@ -37,10 +44,22 @@ export namespace StructurizrBuild {
         async getDependencies(filePath: string, relativePath: string, fileContent: string): Promise<string[]> {
             const dependencies: string[] = [];
 
-            const regex = /workspace\s*extends\s*(.*)\s*{/g;
-            const matches = fileContent.matchAll(regex);
-
+            let regex = /workspace\s*extends\s*(.*)\s*{/g;
+            let matches = fileContent.matchAll(regex);
             let match: IteratorResult<RegExpMatchArray, any>;
+
+            while ((match = matches.next()).done !== true) {
+                const path = this.fileSystem.clearPath(this.fileSystem.getDirectory(filePath), match.value[1].trim());
+
+                if (dependencies.indexOf(path) !== -1) {
+                    continue;
+                }
+
+                dependencies.push(path);
+            }
+
+            regex = /!include\s*([^\n]+)/g;
+            matches = fileContent.matchAll(regex);
 
             while ((match = matches.next()).done !== true) {
                 const path = this.fileSystem.clearPath(this.fileSystem.getDirectory(filePath), match.value[1].trim());
@@ -70,16 +89,34 @@ export namespace StructurizrBuild {
         ) {}
 
         async build(context: Build.FileBuildContext): Promise<void> {
+            if (context.path.endsWith(".extension.dsl")) {
+                const dslFilePath = this.fileSystem.clearPath(
+                    this.buildOptions.outputDirectoryPath,
+                    context.relativePath
+                );
+
+                await this.fileSystem.createDirectory(this.fileSystem.getDirectory(dslFilePath));
+
+                await this.fileSystem.copy(context.path, dslFilePath);
+
+                return;
+            }
+
             const outputDirectoryPath = this.fileSystem.clearPath(
                 this.buildOptions.outputDirectoryPath,
-                context.relativePath
+                context.relativePath + "_processed"
             );
 
             await this.fileSystem.delete(outputDirectoryPath);
 
             this.logger.trace("Rendering PlantUML diagrams with Structurizr CLI");
 
-            await this.structurizrTool.generateDiagrams(context.path, outputDirectoryPath);
+            let dslFileContent = await this.fileSystem.readFile(context.path);
+            dslFileContent = this.transformTemplates(dslFileContent);
+            const dslFilePath = this.fileSystem.clearPath(this.buildOptions.outputDirectoryPath, context.relativePath);
+            await this.fileSystem.createOrOverwriteFile(dslFilePath, dslFileContent);
+
+            await this.structurizrTool.generateDiagrams(dslFilePath, outputDirectoryPath);
 
             const resultFiles = await this.fileSystem.getFilesRecursively(outputDirectoryPath);
 
@@ -101,6 +138,58 @@ export namespace StructurizrBuild {
                 await this.fileSystem.createOrOverwriteFile(outputFilePath, svg);
                 await this.fileSystem.delete(filePath);
             }
+        }
+
+        transformTemplates(fileContent: string): string {
+            const fragments: string[] = [];
+            const templateStartMatches = matchAll(fileContent, templateBeginRegexp);
+            let lastIndex = 0;
+
+            for (let i = 0; i < templateStartMatches.length; i++) {
+                const templateStartMatch = templateStartMatches[i];
+                const startIndex = templateStartMatch.index!;
+                const fragment = fileContent.substring(startIndex);
+
+                fragments.push(fileContent.substring(lastIndex, startIndex));
+
+                const templateBodyMatch = fragment.match(templateBodyRegexp);
+                if (!templateBodyMatch) {
+                    throw new Error("Template has no body tag");
+                }
+                const bodyIndex = startIndex + templateBodyMatch.index! + templateBodyMatch[0].length;
+
+                const templateEndMatch = fragment.match(templateEndRegexp);
+                if (!templateEndMatch) {
+                    throw new Error("Template has no end tag.");
+                }
+                const endIndex = startIndex + templateEndMatch.index!;
+                lastIndex = endIndex + templateEndMatch[0].length;
+
+                const headerFragment = fileContent.substring(startIndex, bodyIndex);
+                const bodyFragment = fileContent.substring(bodyIndex, endIndex);
+
+                matchAll(headerFragment, templateRenderRegexp).forEach((templateRenderMatch) => {
+                    let newFragment = bodyFragment;
+
+                    matchAll(templateRenderMatch[0], renderParametersRegexp).forEach((renderParameterMatch) => {
+                        const parameterName = renderParameterMatch[1];
+                        const parameterValue = renderParameterMatch[2];
+
+                        newFragment = newFragment.replaceAll(`__${parameterName}__`, parameterValue);
+
+                        newFragment = newFragment.replaceAll(
+                            new RegExp(`!constant\\s*${parameterName}\\s*[^\\n]+`, "gm"),
+                            `!constant ${parameterName} ${parameterValue}`
+                        );
+                    });
+
+                    fragments.push(newFragment);
+                });
+            }
+
+            fragments.push(fileContent.substring(lastIndex));
+
+            return fragments.join("\n");
         }
     }
 }
